+++ src/Common/Allocator_fwd.h
@@ -6,5 +6,5 @@
 template <bool clear_memory_, bool mmap_populate = false>
 class Allocator;
 
-template <typename Base, size_t N = 64, size_t Alignment = 1>
+template <typename Base, unsigned long N = 64, unsigned long Alignment = 1>
 class AllocatorWithStackMemory;
