+++ src/Common/ArenaUtils.h
@@ -6,12 +6,12 @@
 
 /** Copy string value into Arena.
   * Arena should support method:
-  * char * alloc(size_t size).
+  * char * alloc(unsigned long size).
   */
 template <typename Arena>
 inline StringRef copyStringInArena(Arena & arena, StringRef value)
 {
-    size_t value_size = value.size;
+    unsigned long value_size = value.size;
     char * place_for_key = arena.alloc(value_size);
     memcpy(reinterpret_cast<void *>(place_for_key), reinterpret_cast<const void *>(value.data), value_size);
     StringRef result{place_for_key, value_size};
